Move SimpleDataPicker to a new directory preparing for new components
