+++ enterprise/frontend/src/embedding-sdk/components/private/data-picker/SimpleDataPicker/SimpleDataPicker.tsx
