+++ enterprise/frontend/src/embedding-sdk/components/private/data-picker/SimpleDataPicker/SimpleDataPickerView.tsx
