+++ enterprise/frontend/src/embedding-sdk/components/private/data-picker/SimpleDataPicker/index.ts
